Clean up dependencies
Rather than placing any requirements on the peer dependencies they are
now left completely open. This means that users may try to use
incompatible versions but in most cases the API dependency is minimal
and just the existence of the package is sufficient.

Additionally, the peer dependencies are no longer used verbatim as the
dev dependencies. Instead specific version ranges of the required
packages (just `coffeescript` and `yaml-js`) are used.